+++ lib/src/actual_phone_state.dart
@@ -2,12 +2,10 @@ import 'package:flutter/services.dart';
 import 'utils/actual_phone_state_status.dart';
 
 class ActualPhoneState {
-  static const MethodChannel _channel =
-      MethodChannel('actual_phone_state');
+  static const MethodChannel _channel = MethodChannel('actual_phone_state');
 
   static Future<ActualPhoneStateStatus> getPhoneState() async {
-    final String? result =
-        await _channel.invokeMethod<String>('getPhoneState');
+    final String? result = await _channel.invokeMethod<String>('getPhoneState');
 
     switch (result) {
       case 'callActive':